+++ lib/excel_plus.dart
@@ -28,6 +28,7 @@ part 'src/models/theme_color.dart';
 part 'src/models/hyperlink.dart';
 part 'src/models/data_validation.dart';
 part 'src/models/sheet_protection.dart';
+part 'src/models/defined_name.dart';
 part 'src/models/enums.dart';
 part 'src/models/span.dart';
 part 'src/models/shared_string.dart';